+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+## NEXT_VERSION
+
+* Change status `closed` to `private` db migration (31d9cf7fb54d)
+
 ## v2.8.1 (2026-08-21)
+
 * Add status dropdown to assessment
 
 ## v2.8.0 (2026-07-21)